# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Anyway Possible Agent Decision APIs",
    "version": "10.0.0",
    "description": "Paid, account-free APIs and MCP tools for x402 merchant intelligence, Base USDC payment readiness, citation evidence, and batch URL validation. Existing routes and prices remain stable."
  },
  "servers": [{ "url": "https://anywaypossible.com" }],
  "paths": {
...
    "/api/treasury": {
      "get": {
        "summary": "Read $0.02 Base USDC payment-preflight metadata",
        "responses": { "200": { "description": "Service metadata" } }
      },
      "post": {
        "operationId": "preflightBaseUsdcPayment",
        "summary": "Preflight a Base USDC payment for $0.02 USDC",
        "description": "Checks live funding and gas, confirms Base chain intent, classifies the destination, catches basic payment hazards, and returns a machine-readable proceed, fund, review, or reject decision.",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TreasuryRequest" } } }
        },
        "responses": {
          "200": { "description": "Paid payment-preflight result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TreasuryResult" } } } },
          "400": { "description": "Invalid address or decimal threshold" },
          "402": { "description": "x402 payment required" },
          "502": { "description": "Base RPC temporarily unavailable" }
        }
      }
    },
    "/api/base-balance": {
      "get": {
        "summary": "Read $0.001 Base wallet balance metadata",
        "responses": { "200": { "description": "Service metadata" } }
      },
      "post": {
        "operationId": "getBaseWalletBalance",
        "summary": "Read live Base ETH and USDC balances for $0.001 USDC",
        "description": "Returns an HTTP 402 x402 challenge when unpaid, then reads native ETH, Circle USDC, and current block height from Base mainnet.",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BaseBalanceRequest" } } }
        },
        "responses": {
          "200": { "description": "Paid Base wallet balance result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BaseBalanceResult" } } } },
          "400": { "description": "Invalid EVM address" },
          "402": { "description": "x402 payment required" },
          "502": { "description": "Base RPC temporarily unavailable" }
        }
      }
    },
    "/api/check": {
      "get": {
        "summary": "Read $0.001 check metadata",
        "responses": { "200": { "description": "Service metadata" } }
      },
      "post": {
        "operationId": "checkUrl",
        "summary": "Preflight a public URL for $0.001 USDC",
        "description": "Returns an HTTP 402 x402 challenge when unpaid, then checks reachability, status, latency, redirects, and content type.",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CheckRequest" } } }
        },
        "responses": {
          "200": { "description": "Paid preflight result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CheckResult" } } } },
          "400": { "description": "Invalid request or unsupported URL" },
          "402": { "description": "x402 payment required" },
          "502": { "description": "The target could not be checked" }
        }
      }
    },
    "/api/batch": {
      "get": {
        "summary": "Read batch check metadata",
        "responses": { "200": { "description": "Service metadata" } }
      },
      "post": {
        "operationId": "batchCheckUrls",
        "summary": "Check up to 10 public URLs for $0.01 USDC",
        "description": "Checks citations, links, or API dependencies in one x402 payment and isolates partial failures per result.",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BatchRequest" } } }
        },
        "responses": {
          "200": { "description": "Paid batch result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BatchResult" } } } },
          "400": { "description": "Invalid request" },
          "402": { "description": "x402 payment required" }
        }
      }
    },
    "/api/verify": {
      "get": {
        "summary": "Read service and payment metadata",
        "responses": {
          "200": {
            "description": "Service metadata",
            "content": { "application/json": { "schema": { "type": "object" } } }
          }
        }
      },
      "post": {
        "operationId": "verifyUrl",
        "summary": "Verify a public URL",
        "description": "Returns an HTTP 402 x402 challenge when unpaid. After a valid $0.01 USDC payment on Base, fetches the URL and returns timestamped verification evidence.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/VerificationRequest" },
              "examples": {
                "basic": {
                  "value": {
                    "url": "https://example.com",
                    "expectedStatus": 200,
                    "expectedText": "Example Domain"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Paid verification result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VerificationResult" }
              }
            }
          },
          "400": { "description": "Invalid request or unsupported URL" },
          "402": {
            "description": "x402 payment required; follow the returned payment requirements and repeat the request with a payment signature"
          },
          "502": { "description": "The target could not be verified" },
          "503": { "description": "Payment service temporarily unavailable" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TreasuryRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["address"],
        "properties": {
          "address": { "type": "string", "pattern": "^0x[a-fA-F0-9]{40}$" },
          "destinationAddress": { "type": "string", "pattern": "^0x[a-fA-F0-9]{40}$", "description": "Optional intended payment recipient; enables destination hazard checks" },
          "plannedSpendUsdc": { "type": "string", "pattern": "^[0-9]+(\\.[0-9]{1,6})?$", "default": "1.00" },
          "minGasReserveEth": { "type": "string", "pattern": "^[0-9]+(\\.[0-9]{1,18})?$", "default": "0.00005" },
          "expectedChainId": { "type": "integer", "enum": [8453], "default": 8453 }
        }
      },
      "TreasuryResult": {
        "type": "object",
        "required": ["address", "destinationAddress", "destinationKind", "network", "chainId", "status", "ready", "safeToProceed", "decision", "riskLevel", "recommendedAction", "eth", "usdc", "plannedSpendUsdc", "minGasReserveEth", "usdcShortfall", "gasShortfallEth", "paymentCapacity", "checks", "alerts", "limitations", "blockNumber", "observedAt"],
        "properties": {
          "address": { "type": "string" },
          "destinationAddress": { "type": ["string", "null"] },
          "destinationKind": { "type": "string", "enum": ["not_provided", "eoa", "contract"] },
          "network": { "type": "string" },
          "chainId": { "type": "integer" },
          "status": { "type": "string", "enum": ["ready", "needs_usdc", "needs_gas", "needs_funding"] },
          "ready": { "type": "boolean" },
          "safeToProceed": { "type": "boolean" },
          "decision": { "type": "string", "enum": ["safe_to_pay", "needs_funding", "needs_gas", "review_destination", "reject"] },
          "riskLevel": { "type": "string", "enum": ["low", "medium", "high"] },
          "recommendedAction": { "type": "string" },
          "eth": { "type": "string" },
          "usdc": { "type": "string" },
          "plannedSpendUsdc": { "type": "string" },
          "minGasReserveEth": { "type": "string" },
          "usdcShortfall": { "type": "string" },
          "gasShortfallEth": { "type": "string" },
          "paymentCapacity": { "type": "object", "additionalProperties": { "type": "string" } },
          "checks": { "type": "array", "items": { "type": "object" } },
          "alerts": { "type": "array", "items": { "type": "string" } },
          "limitations": { "type": "array", "items": { "type": "string" } },
          "usdcContract": { "type": "string" },
          "blockNumber": { "type": "integer" },
          "observedAt": { "type": "string", "format": "date-time" }
        }
      },
      "BaseBalanceRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["address"],
        "properties": { "address": { "type": "string", "pattern": "^0x[a-fA-F0-9]{40}$" } }
      },
      "BaseBalanceResult": {
        "type": "object",
        "required": ["address", "network", "chainId", "eth", "usdc", "ethAtomic", "usdcAtomic", "blockNumber", "observedAt"],
        "properties": {
          "address": { "type": "string" },
          "network": { "type": "string" },
          "chainId": { "type": "integer" },
          "eth": { "type": "string" },
          "usdc": { "type": "string" },
          "ethAtomic": { "type": "string" },
          "usdcAtomic": { "type": "string" },
          "usdcContract": { "type": "string" },
          "blockNumber": { "type": "integer" },
          "observedAt": { "type": "string", "format": "date-time" }
        }
      },
      "VerificationRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["url"],
        "properties": {
          "url": { "type": "string", "format": "uri", "description": "Public HTTP(S) URL to verify" },
          "expectedStatus": { "type": "integer", "minimum": 100, "maximum": 599 },
          "expectedText": { "type": "string", "maxLength": 500 }
        }
      },
      "CheckRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["url"],
        "properties": {
          "url": { "type": "string", "format": "uri", "description": "Public HTTP(S) URL to check" },
          "expectedStatus": { "type": "integer", "minimum": 100, "maximum": 599 }
        }
      },
      "CheckResult": {
        "type": "object",
        "required": ["reachable", "verified", "status", "finalUrl", "responseTimeMs", "observedAt"],
        "properties": {
          "reachable": { "type": "boolean" },
          "verified": { "type": "boolean" },
          "status": { "type": "integer" },
          "finalUrl": { "type": "string", "format": "uri" },
          "redirectChain": { "type": "array", "items": { "type": "string", "format": "uri" } },
          "responseTimeMs": { "type": "integer" },
          "contentType": { "type": "string" },
          "observedAt": { "type": "string", "format": "date-time" }
        }
      },
      "BatchRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["urls"],
        "properties": {
          "urls": { "type": "array", "minItems": 1, "maxItems": 10, "items": { "type": "string", "format": "uri" } },
          "expectedStatus": { "type": "integer", "minimum": 100, "maximum": 599 }
        }
      },
      "BatchResult": {
        "type": "object",
        "required": ["verified", "count", "checkedAt", "results"],
        "properties": {
          "verified": { "type": "boolean" },
          "count": { "type": "integer" },
          "checkedAt": { "type": "string", "format": "date-time" },
          "results": { "type": "array", "items": { "type": "object" } }
        }
      },
      "VerificationResult": {
        "type": "object",
        "required": ["verified", "status", "finalUrl", "contentSha256", "receiptId", "observedAt"],
        "properties": {
          "verified": { "type": "boolean" },
          "status": { "type": "integer" },
          "title": { "type": ["string", "null"] },
          "finalUrl": { "type": "string", "format": "uri" },
          "contentSha256": { "type": "string", "description": "SHA-256 digest of the inspected response bytes" },
          "receiptId": { "type": "string", "description": "Deterministic digest binding URL, status, content digest, and observation time" },
          "redirectChain": { "type": "array", "items": { "type": "string", "format": "uri" } },
          "responseTimeMs": { "type": "integer" },
          "observedAt": { "type": "string", "format": "date-time" }
        }
      }
    }
  }
}